[McpBundle] Extract HTTP middleware resolution into a factory
Move the DNS rebinding protection middleware resolution out of McpController
into a dedicated MiddlewareFactory. The controller no longer decides the
middleware stack; it delegates to the injected factory and stays focused on
Request<->PSR translation. The factory is a plain, dependency-free unit, so
the resolution logic is now tested directly instead of through reflection on
the controller's private method.
