+++ src/mcp-bundle/src/Controller/McpController.php
@@ -12,12 +12,11 @@
 namespace Symfony\AI\McpBundle\Controller;
 
 use Mcp\Server;
-use Mcp\Server\Transport\Http\Middleware\DnsRebindingProtectionMiddleware;
 use Mcp\Server\Transport\StreamableHttpTransport;
 use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\StreamFactoryInterface;
-use Psr\Http\Server\MiddlewareInterface;
 use Psr\Log\LoggerInterface;
+use Symfony\AI\McpBundle\Http\MiddlewareFactory;
 use Symfony\Bridge\PsrHttpMessage\HttpFoundationFactoryInterface;
 use Symfony\Bridge\PsrHttpMessage\HttpMessageFactoryInterface;
 use Symfony\Component\HttpFoundation\Request;
@@ -25,19 +24,14 @@ use Symfony\Component\HttpFoundation\Response;
 
 final class McpController
 {
-    /**
-     * @param list<string>|false|null $allowedHosts Hostnames allowed by the DNS rebinding protection: null keeps the
-     *                                              SDK default (localhost only), an array restricts it to those hosts,
-     *                                              false disables the protection entirely
-     */
     public function __construct(
         private readonly Server $server,
         private readonly HttpMessageFactoryInterface $httpMessageFactory,
         private readonly HttpFoundationFactoryInterface $httpFoundationFactory,
         private readonly ResponseFactoryInterface $responseFactory,
         private readonly StreamFactoryInterface $streamFactory,
+        private readonly MiddlewareFactory $middlewareFactory,
         private readonly ?LoggerInterface $logger = null,
-        private readonly array|false|null $allowedHosts = null,
     ) {
     }
 
@@ -48,43 +42,12 @@ final class McpController
             $this->responseFactory,
             $this->streamFactory,
             logger: $this->logger,
-            middleware: $this->resolveMiddleware(),
+            middleware: $this->middlewareFactory->create(),
         );
 
         $psrResponse = $this->server->run($transport);
         $streamed = 'text/event-stream' === strtolower($psrResponse->getHeaderLine('Content-Type'));
 
         return $this->httpFoundationFactory->createResponse($psrResponse, $streamed);
-    }
-
-    /**
-     * Returns null to keep the SDK secure defaults, or a tailored middleware stack when the DNS
-     * rebinding protection needs to be disabled or restricted to custom hosts.
-     *
-     * @return list<MiddlewareInterface>|null
-     */
-    private function resolveMiddleware(): ?array
-    {
-        // Unset: keep the SDK secure defaults (DNS rebinding protection limited to localhost).
-        if (null === $this->allowedHosts) {
-            return null;
-        }
-
-        $middleware = [];
-        foreach (StreamableHttpTransport::defaultMiddleware() as $defaultMiddleware) {
-            if ($defaultMiddleware instanceof DnsRebindingProtectionMiddleware) {
-                // false: disable the protection entirely to expose a public MCP server.
-                if (false === $this->allowedHosts) {
-                    continue;
-                }
-
-                // array: restrict the protection to the configured hosts.
-                $defaultMiddleware = new DnsRebindingProtectionMiddleware($this->allowedHosts);
-            }
-
-            $middleware[] = $defaultMiddleware;
-        }
-
-        return $middleware;
     }
 }
